Let chromium be able to send stream data to QUIC via WriteMemSlices. Protected by --quic_reloadable_flag_quic_use_mem_slices.
R=rch@chromium.org

Bug:

+++ net/quic/platform/impl/quic_test_mem_slice_vector_impl.h
@@ -22,14 +22,14 @@ class QUIC_EXPORT_PRIVATE TestIOBuffer : public IOBuffer {
 class QUIC_EXPORT_PRIVATE QuicTestMemSliceVectorImpl {
  public:
   explicit QuicTestMemSliceVectorImpl(
-      std::vector<std::pair<char*, size_t>> buffers);
+      std::vector<std::pair<char*, int>> buffers);
   ~QuicTestMemSliceVectorImpl();
 
   QuicMemSliceSpanImpl span();
 
  private:
   std::vector<scoped_refptr<IOBuffer>> buffers_;
-  std::vector<size_t> lengths_;
+  std::vector<int> lengths_;
 };
 
 }  // namespace test
